# Use VTproto generated clone method