@luluz66 2024-01-05
> Is there a way for you to rebase so that this doesn't have a bunch of extra commit history in the PR?

Done. `git push --force` is the way to go.